Move sync-related items to new sync app.
Only sync-related items not moved are the VCS functions in
pontoon.administration.vcs since the old sync still relies
on them.
