+++ pontoon/base/__init__.py
@@ -14,7 +14,3 @@ MOZILLA_REPOS = (
     'ssh://hg.mozilla.org/users/m_owca.info/lightning-aurora/',
     'ssh://hg.mozilla.org/users/m_owca.info/seamonkey-aurora/',
 )
-
-
-class SyncError(RuntimeError):
-    """Error class for errors relating to the project sync process."""
