+++ pontoon/sync/__init__.py
@@ -1,0 +1,2 @@
+class SyncError(RuntimeError):
+    """Error class for errors relating to the project sync process."""
