+++ pontoon/sync/formats/xliff.py
@@ -3,8 +3,8 @@ Parser for the xliff translation format.
 """
 from translate.storage import xliff
 
-from pontoon.base.formats.base import ParsedResource
-from pontoon.base.vcs_models import VCSTranslation
+from pontoon.sync.formats.base import ParsedResource
+from pontoon.sync.vcs_models import VCSTranslation
 
 
 class XLIFFEntity(VCSTranslation):
