+++ pontoon/sync/tests/directory_detection_tests/no_resources_test/real_resources/templates/test.po
