+++ pontoon/sync/tests/directory_detection_tests/no_resources_test/templates/not_a_resource.py
