+++ pontoon/sync/tests/directory_detection_tests/scoring_en_US_test/en-US/test.po
