+++ pontoon/sync/tests/directory_detection_tests/scoring_source_files_test/en/test.pot
