+++ pontoon/sync/tests/directory_detection_tests/scoring_source_files_test/templates/test.po
