+++ pontoon/sync/tests/directory_detection_tests/scoring_templates_test/en-US/test.po
