+++ pontoon/sync/tests/fake-checkout/inactive-locale/main.lang
