+++ pontoon/sync/tests/fake-checkout/inactive-locale/other.lang
