+++ pontoon/sync/tests/fake-checkout/templates/main.lang
