+++ pontoon/sync/tests/fake-checkout/templates/other.lang
