+++ pontoon/sync/tests/fake-checkout/translated-locale/main.lang
