+++ pontoon/sync/tests/fake-checkout/translated-locale/other.lang
