+++ pontoon/sync/tests/formats/bom.lang
